pyproject.toml
[project]
name = "epic-fhir-mcp-server"
version = "0.1.0"
description = "MCP Server for Epic FHIR API integration"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.1.0",
    "pyjwt[crypto]>=2.8.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.37.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    "cryptography>=42.0.0",
    "openai-agents>=0.3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
agent = [
    "openai-agents>=0.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
